# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProviEmplea API",
        "description": "API REST para la plataforma ProviEmplea.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/personas": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Listar personas",
                "description": "Retorna el listado de personas registradas en ProviEmplea.",
                "operationId": "84e537aeba1e2d52163ea10927e98430",
                "responses": {
                    "200": {
                        "description": "Listado de personas obtenido correctamente"
                    }
                }
            },
            "post": {
                "tags": [
                    "Personas"
                ],
                "summary": "Crear persona",
                "description": "Registra una nueva persona en ProviEmplea.",
                "operationId": "9d6d475142411d09ace30132072dbeb7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "email": "talento@ejemplo.cl",
                                "telefono": "+56912345678",
                                "codigo_talento": "PROV002",
                                "nivel_educacional": "Universitaria",
                                "titulo_carrera": "Ingeniería Informática",
                                "anio_egreso": 2024,
                                "anios_experiencia": 1,
                                "competencias": [
                                    "PHP",
                                    "Laravel"
                                ],
                                "areas_experiencia": [
                                    "Backend"
                                ],
                                "rango_renta": "800000",
                                "tipo_jornada": "Completa",
                                "modalidad": "Híbrida",
                                "cursos": [
                                    "Laravel Básico"
                                ],
                                "idiomas": [
                                    "Español"
                                ],
                                "persona_discapacidad": false,
                                "validado": false,
                                "activo": true
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Persona creada correctamente"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/personas/{id}": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener persona por ID",
                "description": "Retorna una persona específica según su ID.",
                "operationId": "bbd0a2bc22755dc3676325845d14deb2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona encontrada"
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Personas"
                ],
                "summary": "Actualizar persona",
                "description": "Actualiza los datos de una persona existente.",
                "operationId": "a93de69e16ad096453410414ed502809",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "telefono": "+56987654321",
                                "nivel_educacional": "Técnica",
                                "anios_experiencia": 2,
                                "modalidad": "Remota"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Persona actualizada correctamente"
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Personas"
                ],
                "summary": "Eliminar persona",
                "description": "Elimina una persona registrada según su ID.",
                "operationId": "42bc19ebb4f9057d921a5d3fad327b28",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona eliminada correctamente"
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/personas/{id}/validar": {
            "patch": {
                "tags": [
                    "Personas"
                ],
                "summary": "Validar persona",
                "description": "Marca una persona como validada.",
                "operationId": "25e6ef5f6294d18ff84fd12bde5167b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona validada correctamente"
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                }
            }
        }
    },
...
}